refactor: (#11) ChatBox 이름 수정 및 index.tsx 작성, 구조 리팩터링
