--- src/features/main/ui/Chat/index.ts
@@ -1,7 +1,0 @@
-export { ChatBox } from './ChatBox';
-export { ChannelSelect } from './ChannelSelect';
-export { ChatInput } from './ChatInput';
-export { ChatInputContainer } from './ChatInputContainer';
-export { ChatSendButton } from './ChatSendButton';
-export { MentionDropdown } from './MentionDropdown';
-export { MessageList } from './MessageList';
